+++ src/client/components/edit_btn/edit_btn.js
@@ -1,10 +1,12 @@
 /*
  * pwix:tabular/src/client/components/edit_btn/edit_btn.js
  *
- * Data context is provided at the constructor level:
+ * Data context is provided at instanciation time:
  * - item: the row data
  * - table: the Tabular.Table instance
  */
+
+import _ from 'lodash';
 
 import { pwixI18n } from 'meteor/pwix:i18n';
 import { ReactiveVar } from 'meteor/reactive-var';
@@ -15,36 +17,26 @@ Template.edit_btn.onCreated( function(){
     const self = this;
 
     self.PCK = {
-        item: new ReactiveVar( null ),
+        // have a deep copy of provided data context
+        dataContext: new ReactiveVar( null ),
+        // manage the UI
         enabled: new ReactiveVar( true ),
         title: new ReactiveVar( '' )
     };
-});
 
-Template.edit_btn.onRendered( function(){
-    const self = this;
-
-    // get asynchronously the item
-    self.autorun(() => {
-        const dc = Template.currentData();
-        if( dc.item && dc.table ){
-            dc.table.opt( 'editItem', dc.item, dc.item ).then(( res ) => { self.PCK.item.set( res ); });
-       }
-    });
-
-    // get asynchronously the enabled state
-    self.autorun(() => {
-        const dc = Template.currentData();
-        if( dc.item && dc.table ){
-            dc.table.opt( 'editButtonEnabled', true, dc.item ).then(( res ) => { self.PCK.enabled.set( res ); });
-        }
-    });
-
-    // get asynchronously the button title
-    self.autorun(() => {
-        const dc = Template.currentData();
-        if( dc.item && dc.table ){
-            dc.table.opt( 'editButtonTitle', pwixI18n.label( I18N, 'edit.btn_title', dc.item._id ), dc.item ).then(( res ) => { self.PCK.title.set( res ); });
+    // keep a copy of the initial data context
+    self.autorun(( comp ) => {
+        const dataContext = Template.currentData();
+        if( dataContext.item && dataContext.table ){
+            self.PCK.dataContext.set( _.cloneDeep( dataContext ));
+            // table.opt() prototype is ( name, default, row_value )
+            dataContext.table.opt( 'editItem', dataContext.item, dataContext.item );
+            dataContext.table.opt( 'editButtonEnabled', true, dataContext.item )
+                .then(( res ) => { self.PCK.enabled.set( res ); })
+            dataContext.table.opt( 'editButtonTitle', pwixI18n.label( I18N, 'edit.btn_title', dataContext.item._id ), dataContext.item )
+                .then(( res ) => { self.PCK.title.set( res ); });
+            // run only once
+            comp.stop();
         }
     });
 });
@@ -55,7 +47,7 @@ Template.edit_btn.helpers({
         return Template.instance().PCK.enabled.get() || !Tabular.configure().hideDisabled ? '' : 'ui-transparent';
     },
 
-    // whether the displayed row is deletable ? defaulting to true
+    // whether the displayed row is editable ? defaulting to true
     enabledState(){
         return Template.instance().PCK.enabled.get() ? '' : 'disabled';
     },
@@ -68,8 +60,6 @@ Template.edit_btn.helpers({
 
 Template.edit_btn.events({
     async 'click .tabular-edit-btn button'( event, instance ){
-        const item = instance.PCK.item.get();
-        this.item = item;
-        instance.$( event.currentTarget ).trigger( 'tabular-edit-event', this );
+        instance.$( event.currentTarget ).trigger( 'tabular-edit-event', instance.PCK.dataContext.get());
     }
 });