+++ src/client/components/info_btn/info_btn.js
@@ -1,6 +1,12 @@
 /*
  * pwix:tabular/src/client/components/info_btn/info_btn.js
+ *
+ * Data context is provided at instanciation time:
+ * - item: the row data
+ * - table: the Tabular.Table instance
  */
+
+import _ from 'lodash';
 
 import { pwixI18n } from 'meteor/pwix:i18n';
 import { ReactiveVar } from 'meteor/reactive-var';
@@ -11,45 +17,26 @@ Template.info_btn.onCreated( function(){
     const self = this;
 
     self.PCK = {
-        object: new ReactiveVar( null ),
+        // have a deep copy of provided data context
+        dataContext: new ReactiveVar( null ),
+        // manage the UI
         enabled: new ReactiveVar( true ),
         title: new ReactiveVar( '' )
     };
 
-    // track the target object which may come from an async function
-    self.autorun(() => {
+    // keep a copy of the initial data context
+    self.autorun(( comp ) => {
         const dataContext = Template.currentData();
-        let object = dataContext.item;
-        let promises = [];
-        const p = dataContext.table.opt( 'infoItem', dataContext.item, dataContext.item );
-        if( p instanceof Promise ){
-            promises.push( p.then(( res ) => {
-                object = res;
-                return res;
-            }));
-        }
-        Promise.allSettled( promises ).then(() => {
-            self.PCK.object.set( object );
-        });
-    });
-});
-
-Template.info_btn.onRendered( function(){
-    const self = this;
-
-    // get asynchronously the enabled state
-    self.autorun(() => {
-        const dc = Template.currentData();
-        if( dc.item && dc.table ){
-            dc.table.opt( 'infoButtonEnabled', true, dc.item ).then(( res ) => { self.PCK.enabled.set( res ); });
-        }
-    });
-
-    // get asynchronously the button title
-    self.autorun(() => {
-        const dc = Template.currentData();
-        if( dc.item && dc.table ){
-            dc.table.opt( 'infoButtonTitle', pwixI18n.label( I18N, 'info.btn_title', dc.item._id ), dc.item ).then(( res ) => { self.PCK.title.set( res ); });
+        if( dataContext.item && dataContext.table ){
+            self.PCK.dataContext.set( _.cloneDeep( dataContext ));
+            // table.opt() prototype is ( name, default, row_value )
+            dataContext.table.opt( 'infoItem', dataContext.item, dataContext.item );
+            dataContext.table.opt( 'infoButtonEnabled', true, dataContext.item )
+                .then(( res ) => { self.PCK.enabled.set( res ); })
+            dataContext.table.opt( 'infoButtonTitle', pwixI18n.label( I18N, 'info.btn_title', dataContext.item._id ), dataContext.item )
+                .then(( res ) => { self.PCK.title.set( res ); });
+            // run only once
+            comp.stop();
         }
     });
 });
@@ -58,11 +45,12 @@ Template.info_btn.helpers({
 
     // miButton parameters
     parmsModalInfo(){
+        const dataContext = Template.instance().PCK.dataContext.get();
         return {
             titleButton: Template.instance().PCK.title.get(),
-            titleDialog: this.table.opt( 'infoModalTitle', pwixI18n.label( I18N, 'modalinfo.dialog_title', this.item._id ), this.item ),
-            mdClassesContent: this.table.opt( 'dialogClasses', '', this.item ),
-            object: Template.instance().PCK.object.get(),
+            titleDialog: this.table.opt( 'infoModalTitle', pwixI18n.label( I18N, 'modalinfo.dialog_title', dataContext.item._id ), dataContext.item ),
+            mdClassesContent: this.table.opt( 'dialogClasses', '', dataContext.item ),
+            object: dataContext.item,
             classButton: 'btn-sm btn-outline-primary',
             stampFormat: '%Y-%m-%d %H:%M:%S',
             enabled: Template.instance().PCK.enabled.get()
@@ -77,6 +65,6 @@ Template.info_btn.helpers({
 
 Template.info_btn.events({
     'click .tabular-info-btn button'( event, instance ){
-        instance.$( event.currentTarget ).trigger( 'tabular-info-event', this );
+        instance.$( event.currentTarget ).trigger( 'tabular-info-event', instance.PCK.dataContext.get());
     }
 });